chore(dev.yml): update dev.yml workflow to include separate docker job for building Docker images chore(docker.yml): remove docker.yml workflow file as it is no longer needed
The dev.yml workflow file was updated to include a separate job named 'docker' for building Docker images. This separation allows for better organization and clarity in the workflow process. The docker.yml workflow file was removed as its functionality is now covered by the new 'docker' job in dev.yml.