feat(ssm-bdd): add InvokeChunkedProps class to support chunking of data processing in Cucumber tests feat(ssm-chaincode-dsl): introduce chunking property in SsmChaincodePropertiesDTO interface to configure chunking behavior refactor(ssm-sdk-core): remove unused imports and classes related to chaincode URI and signing test(ssm-sdk-core): update SsmClientArrayTest to use SsmTxService instead of SsmCommandService for transaction handling feat(ssm-tx-spring-boot): add chunking property to SsmTxProperties to configure chunking behavior for transactions
The changes were made to enhance the codebase by introducing a new class, InvokeChunkedProps, to handle chunking of data processing in Cucumber tests. Additionally, the chunking property was added to the SsmChaincodePropertiesDTO interface and SsmTxProperties class to allow configuration of chunking behavior. Unused imports and classes related to chaincode URI and signing were removed for code cleanliness. The SsmClientArrayTest was updated to use SsmTxService instead of SsmCommandService for transaction handling.

feat(ssm-tx): introduce SsmTxProperties to autoconfigure classes in SsmTx modules

The changes involve introducing a new SsmTxProperties class to autoconfigure
various classes in the SsmTx modules. This allows for better configuration
management and enhances the flexibility of the SsmTx modules by providing
properties that can be injected into beans during autoconfiguration.

feat(ssm-tx): introduce InvokeChunkedProps for chunking operations in SsmTxAdminServiceImpl, SsmTxUserServiceImpl, SsmTxCreateFunctionImpl, and SsmTxSessionPerformActionFunctionImpl

The changes introduce the InvokeChunkedProps class to handle chunking operations in various functions across the SsmTxAdminServiceImpl, SsmTxUserServiceImpl, SsmTxCreateFunctionImpl, and SsmTxSessionPerformActionFunctionImpl classes. This enhancement improves the efficiency and performance of chunking operations within the specified functions.

feat(SsmTxSessionStartFunctionImpl.kt): refactor code to chunk and flatten coroutine flows for improved performance
feat(SsmUserGrantFunctionImpl.kt): refactor code to chunk and flatten coroutine flows for improved performance
feat(SsmUserRegisterFunctionImpl.kt): refactor code to chunk and flatten coroutine flows for improved performance

The changes were made to refactor the code in the mentioned Kotlin files to utilize chunking and flattening of coroutine flows. This optimization enhances the performance of the application by efficiently processing the data in chunks and concurrently flattening the results.